+++ include/metalang99/eval/syntax_checker.h
@@ -14,7 +14,7 @@
 
 // clang-format off
 #define ML99_PRIV_SYNTAX_ERROR(invalid_term) \
-    ML99_PRIV_REC_CONTINUE(ML99_PRIV_REC_STOP)((~), !"Metalang99 syntax error": `invalid_term`)
+    ML99_PRIV_REC_CONTINUE(ML99_PRIV_REC_STOP)((~), !"Metalang99 syntax error": {invalid_term})
 // clang-format on
 
 #endif // ML99_EVAL_SYNTAX_CHECKER_H